requirements.txt
streamlit>=1.36
google-generativeai>=0.8.0
chromadb>=0.5.5
langchain-text-splitters>=0.0.1
pypdf>=4.2.0
python-dotenv>=1.0.1
